refactor: load doubles instead of ints

+++ include/GUI/MapLoader.hpp
@@ -1,10 +1,10 @@
 #pragma once
-#include <Qstring>
+#include <QString>
 #include <map>
 #include <vector>
 
 class MapLoader {
  public:
-  static const std::map<std::string, std::vector<std::map<std::string, int>>>
+  static const std::map<std::string, std::vector<std::map<std::string, double>>>
   loadMap(QString path);
 };
